Turn `:std:uri` into `:ext:uri`
