+++ core/mastodon/build.gradle.kts
@@ -54,13 +54,13 @@ dependencies {
   implementation(project(":composite:timeline"))
   implementation(project(":core:sample"))
   implementation(project(":ext:coroutines"))
+  implementation(project(":ext:uri"))
   implementation(project(":platform:autos"))
   implementation(project(":platform:cache"))
   implementation(project(":platform:focus"))
   implementation(project(":platform:navigation"))
   implementation(project(":platform:starter"))
   implementation(project(":std:visibility"))
-  implementation(project(":std:uri"))
   implementation(libs.android.browser)
   implementation(libs.android.constraintLayout.compose)
   implementation(libs.android.fragment.ktx)
