+++ core/sample/src/main/java/br/com/orcinus/orca/core/sample/feed/profile/post/Post.extensions.kt
@@ -29,10 +29,10 @@ import br.com.orcinus.orca.core.sample.feed.profile.post.stat.createSampleToggle
 import br.com.orcinus.orca.core.sample.image.CoverImageSource
 import br.com.orcinus.orca.core.sample.image.SampleImageSource
 import br.com.orcinus.orca.core.sample.instance.domain.sample
+import br.com.orcinus.orca.ext.uri.URIBuilder
 import br.com.orcinus.orca.std.image.ImageLoader
 import br.com.orcinus.orca.std.image.SomeImageLoaderProvider
 import br.com.orcinus.orca.std.markdown.buildMarkdown
-import br.com.orcinus.orca.std.uri.URIBuilder
 import java.time.ZoneId
 import java.time.ZonedDateTime
 import java.util.UUID
