+++ core/shared-preferences/build.gradle.kts
@@ -40,8 +40,8 @@ dependencies {
   testImplementation(project(":core:sample-test"))
   testImplementation(project(":core-test"))
   testImplementation(project(":ext:reflection"))
+  testImplementation(project(":ext:uri"))
   testImplementation(project(":platform:testing"))
-  testImplementation(project(":std:uri"))
   testImplementation(libs.assertk)
   testImplementation(libs.kotlin.coroutines.test)
   testImplementation(libs.kotlin.test)
