+++ core/src/test/java/br/com/orcinus/orca/core/feed/profile/post/content/ContentTests.kt
@@ -20,9 +20,9 @@ import br.com.orcinus.orca.core.feed.profile.post.content.highlight.Highlight
 import br.com.orcinus.orca.core.instance.domain.Domain
 import br.com.orcinus.orca.core.sample.instance.domain.sample
 import br.com.orcinus.orca.core.sample.test.feed.profile.post.content.highlight.sample
+import br.com.orcinus.orca.ext.uri.url.HostedURLBuilder
 import br.com.orcinus.orca.std.markdown.Markdown
 import br.com.orcinus.orca.std.markdown.buildMarkdown
-import br.com.orcinus.orca.std.uri.url.HostedURLBuilder
 import kotlin.test.Test
 import kotlin.test.assertEquals
