+++ ext/uri/src/test/java/br/com/orcinus/orca/ext/uri/URIBuilderTests.kt
@@ -13,36 +13,23 @@
  * not, see https://www.gnu.org/licenses.
  */
 
-package br.com.orcinus.orca.std.uri.url
+package br.com.orcinus.orca.ext.uri
 
 import assertk.assertThat
 import assertk.assertions.isEqualTo
-import br.com.orcinus.orca.std.uri.URIBuilder
+import br.com.orcinus.orca.ext.uri.url.SchemedURLBuilder
 import java.net.URI
 import kotlin.test.Test
 
-internal class SegmentedURLBuilderTests {
+class URIBuilderTests {
   @Test
-  fun appendsParameters() {
-    assertThat(
-        URIBuilder.url()
-          .scheme("https")
-          .host("mastodon.social")
-          .path("api")
-          .path("v1")
-          .path("statuses")
-          .path("112276588128366269")
-          .path("reblogged_by")
-          .query()
-          .parameter("limit", "2")
-          .parameter("since_id", "112276666465473478")
-          .build()
-      )
-      .isEqualTo(
-        URI(
-          "https://mastodon.social/api/v1/statuses/112276588128366269/reblogged_by?limit=2&since_" +
-            "id=112276666465473478"
-        )
-      )
+  fun createsSchemedURIBuilder() {
+    assertThat(URIBuilder.url().scheme("https")).isEqualTo(SchemedURLBuilder("https"))
+  }
+
+  @Test
+  fun buildsMailtoURI() {
+    assertThat(URIBuilder.mailto("jean@orcinus.com.br"))
+      .isEqualTo(URI("${URIBuilder.MAILTO}:jean@orcinus.com.br"))
   }
 }
