+++ platform/intents-test/build.gradle.kts
@@ -29,7 +29,7 @@ android {
 dependencies {
   debugImplementation(libs.android.lifecycle.runtime)
 
-  implementation(project(":std:uri"))
+  implementation(project(":ext:uri"))
   implementation(libs.android.test.espresso.intents)
 
   testImplementation(project(":platform:starter"))
