+++ std/markdown/build.gradle.kts
@@ -20,7 +20,7 @@ plugins {
 }
 
 dependencies {
-  testImplementation(project(":std:uri"))
+  testImplementation(project(":ext:uri"))
   testImplementation(libs.assertk)
   testImplementation(libs.kotlin.test)
 }
