--- std/uri/build.gradle.kts
@@ -1,10 +1,0 @@
-plugins {
-  alias(libs.plugins.kotlin.jvm)
-
-  `java-library`
-}
-
-dependencies {
-  testImplementation(libs.assertk)
-  testImplementation(libs.kotlin.test)
-}
